Move headers only used by plugins to plugins folder
