+++ RecoHGCal/TICL/plugins/GlobalCache.h
