+++ RecoHGCal/TICL/plugins/SeedingRegionAlgoBase.h
